requirements.txt
beautifulsoup4==4.12.3
aiohttp==3.10.10
asyncio==3.4.3
fake-useragent==1.5.1
lxml==5.3.0
tqdm==4.66.5
scipy==1.14.1
liac-arff==2.5.0
pandas==2.2.3
numpy==2.1.2
requests==2.32.3